Revert "added trace"
This reverts commit aff0a339ed1051f538b589702517563e283d2e5a.
