+++ .gitignore
@@ -38,4 +38,3 @@ hs_err_*.log
 replay_*.log
 *.hprof
 *.jfr
-.kotlin/errors
